style: add member profile component
