+++ client/collabor8/src/app/profile-edit/page-edit.css
@@ -1,0 +1,8 @@
+.profile {
+    display: flex;
+    flex-direction: column;
+}
+
+.profile__button {
+    width: 200px;
+}
