--- client/collabor8/src/app/profile/style.css
@@ -1,56 +1,0 @@
-.profile {
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    height: 100%;
-}
-
-.profile__container {
-    width: 608px;
-    padding: 31px 26px;
-    border-radius: 24px;
-    background-color: var(--white);
-    box-shadow: 0px 4px 10.6px -3px rgba(0, 0, 0, 0.25);
-}
-
-.profile__title {
-    color: var(--interface-black, #000);
-    font-family: var(--headers-font);
-    font-size: var(--h5-size);   
-}
-
-.profile__input {
-    display: flex;
-    width: 100%;
-    margin-top: 24px;
-}
-.prifile__item {
-    flex: 1;
-}
-
-.right {
-    margin-right: 16px;
-}
-
-.top {
-    margin-top: 24px;
-}
-
-.input__label {
-    color: var(--interface-gray600);
-    margin-left: 6px;
-    margin-bottom: 3px;
-    font-size: var(--body3-size);
-}
-
-.profile__tag {
-    display: flex;
-    margin-top: 10px;
-    overflow: hidden;
-    overflow-x: scroll;
-    max-width: 540px;
-}
-
-.profile__button {
-    margin-top: 40px;
-}
