+++ client/collabor8/src/components/member-profile/member-profile.css
@@ -1,0 +1,20 @@
+.user {
+    display: flex;
+    background-color: white;
+    justify-content: flex-start;
+    /* width: 300px; */
+}
+
+.profile__column{
+    flex-direction: column;
+}
+
+.profile__row{
+    flex-direction: row;
+}
+
+.profile__items {
+    display: flex;
+    flex-direction: column;
+    
+}
